requirements.txt
pytest==6.2.4
requests==2.25.1
loguru==0.5.3
selenium==3.141.0
webdriver-manager==3.4.2
allure-pytest==2.9.43
PyAutoIt==0.6.5
Pillow~=8.3.1
pytest-bdd~=5.0.0
pytest-html~=3.1.1